Fix large tensor index overflow in CUDA kernels

+++ sgl-kernel/csrc/mamba/causal_conv1d.h
@@ -10,7 +10,7 @@
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 
 struct ConvParamsBase {
-    using index_t = uint32_t;
+    using index_t = int64_t;
 
     int batch, dim, seqlen, width;
     int64_t pad_slot_id;
